Fix strange bug with regex. Wrongfully extracted переулок from OSM name Первомайский переулок, despite correctly working on Regex101.com
