+++ src/main/kotlin/org/openstreetmap/josm/plugins/dl/russiaaddresshelper/settings/TagSettingsPanel.kt
@@ -1,6 +1,6 @@
 package org.openstreetmap.josm.plugins.dl.russiaaddresshelper.settings
 
-import org.openstreetmap.josm.plugins.dl.russiaaddresshelper.io.TagSettingsReader
+import org.openstreetmap.josm.plugins.dl.russiaaddresshelper.settings.io.TagSettingsReader
 import org.openstreetmap.josm.tools.GBC
 import org.openstreetmap.josm.tools.I18n
 import java.awt.GridBagLayout
