+++ src/main/resources/references/street_types.yml
@@ -13,7 +13,7 @@ types:
   - name: переулок
     osm:
       patterns:
-        - '^(?:пер\S*\s+)?(?<street>.+?)(?:\s+пер\S*)?$' # переулок Трактористов, Переходный переулок
+        - '^(?:переулок\s?)?(?<street>.+?)(?:\s?переулок)?$' # переулок Трактористов, Переходный переулок
     egrn:
       patterns:
         - '^[^:]+((?:(пер(?:\.|\s))|переулок)\s*)+(?:им\.?\s)?(?<street>[0-9A-Я"]+[а-яA-Я0-9-"\.]*(?:\s(?:[а-яA-Я0-9-"\.])+)*)' # Свердловская обл., г. Екатеринбург, пер. Переходный, 7
